Renamed `VehicleRemoteAuthenticator` to `Types`
